requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.0
httpx==0.27.0
python-docx==1.1.0
docx2pdf==0.1.8
firebase-admin==6.5.0
PyJWT==2.8.0
PyGithub==2.3.0
ollama==0.2.0
toml 
pypdf 
PyYAML 
requests 
